fast-graphql/pyproject.toml
[project]
name = "fast-graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic==1.18.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "bcrypt>=5.0.0",
    "broadcaster[redis]>=0.3.1",
    "click==8.3.1",
    "colorama==0.4.6",
    "cross-web==0.4.1",
    "fastapi==0.128.0",
    "graphql-core==3.2.7",
    "greenlet==3.3.1",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "packaging==26.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "pymysql==1.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "ruff>=0.15.1",
    "six==1.17.0",
    "sqlalchemy==2.0.46",
    "starlette==0.50.0",
    "strawberry-graphql==0.289.8",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
]
[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
